Add Tests For Data Container Batching
